Write database.yml before checkout.
The ideal behaviour would be to wait for the checkout in order to be able to examine Gemfile, application.rb and the like to find out what database adapter is most appropriate. The database config also has to be in place before deploy hooks are triggered, which makes it a bit tricky. These checks will be readded as soon as we have a fix for Chef that prevents Chef from removing generated template files too early.