Allow deploy_to to override application install location
Currently every other reference to /srv/www in the context of
application deployment is configurable. Adjusting the :deploy_to
configuration variable in an application changes the location of the
install. However in this one case in nginx uninstall /srv/www is
hard coded as the directory prefix.

cr https://cr.amazon.com/r/2565517/