Revert "Check for existing Chef client YAML file."
This reverts commit e13af46be0e14756b4e2ed3297f0bdb047e5eac3.

The reverted commit was an attempt to make sure we always load a YAML
config file, no matter under which name it exists in the file system.
However, this might cause hard-to-debug problems because we use config
values for stage2 in a run for stage1. In addition to reverting this
change I'm also making sure that client.yml is a symlink that always
points to the right physical YAML file, matching the stage we're
currently in.

cr https://cr.amazon.com/r/2842282/