Use service command for managing ganglia-monitor on ubuntu 14.04.
Chef uses the debian init resource for ubuntu 14.04 because there are
some issues with the upstart resouce. The ganglia-monitor init script
has no status method. This means that start-stop-service will fail with
an error if the chef run tries to start an already started
ganglia-monitor service.

For this reason we had explicit stop commands and pkill before to get
around the inconsistency when we start ganglia-monitor with service
(then you have to stop it with service - stopping it via init script
will restart it imminently.) and init script (then you have to check if
service is already started before starting it again).

When we manage ganglia-monitor via the init script upstart is not aware
that this service is running. I think using the service wrapper (which
defaults to upstart's start/stop/... scripts if a upstart config exists)
is the best approach in this case.